requirements.txt
fastapi
uvicorn
openai
pinecone
python-dotenv
python-multipart
pypdf
